1) Updated unit tests 2) Fix memory leaks identified by tests run under valgrind

+++ src/test/test4.c
@@ -21,7 +21,6 @@
  *
  */
 
-
 #ifdef RUN_UNIT_TESTS
 
 #define _GNU_SOURCE
@@ -37,9 +36,58 @@
 #include <signal.h>
 #include <syslog.h>
 #include <pthread.h>
+#include <assert.h>
 
 #include "worker-threads.h"
 #include "dragonfly-io.h"
+
+#include "test.h"
+
+static const char *CONFIG_LUA =
+	"inputs = {\n"
+	"   { tag=\"input\", uri=\"ipc://input.ipc\", script=\"input.lua\"}\n"
+	"}\n"
+	"\n"
+	"analyzers = {\n"
+	"    { tag=\"test\", script=\"analyzer.lua\" },\n"
+	"}\n"
+	"\n"
+	"outputs = {\n"
+	"    { tag=\"log\", uri=\"file://test.log\"},\n"
+	"}\n"
+	"\n";
+
+static const char *INPUT_LUA =
+	"function setup()\n"
+	"end\n"
+	"\n"
+	"function loop(msg)\n"
+	"   analyze_event (\"test\", msg)\n"
+	"end\n";
+
+static const char *ANALYZER_LUA =
+	"function setup()\n"
+	"end\n"
+	"function loop (msg)\n"
+	"   output_event (\"log\", msg)\n"
+	"end\n\n";
+/*
+ * ---------------------------------------------------------------------------------------
+ *
+ * ---------------------------------------------------------------------------------------
+ */
+static void write_file(const char *file_path, const char *content)
+{
+	fprintf(stderr, "generated %s\n", file_path);
+	FILE *fp = fopen(file_path, "w+");
+	if (!fp)
+	{
+		perror(__FUNCTION__);
+		return;
+	}
+	fputs(content, fp);
+	fclose(fp);
+}
 
 /*
  * ---------------------------------------------------------------------------------------
@@ -48,72 +96,36 @@
  */
 void SELF_TEST4(const char *dragonfly_root)
 {
+#define MAX_TEST4_MESSAGES 10000
+	int halfway = (MAX_TEST4_MESSAGES / 2);
+	const char *analyzer_path = "./scripts/analyzer.lua";
+	const char *input_path = "./scripts/input.lua";
+	const char *config_path = "./scripts/config.lua";
 
-#define MAX_TEST4_MESSAGES 150000
-	int halfway = (MAX_TEST4_MESSAGES/2);
-	fprintf(stderr, "\n\n%s:(pumping %d messages and rotating at %d)\n", __FUNCTION__, MAX_TEST4_MESSAGES, halfway);
+	fprintf(stderr, "\n\n%s: pumping %d messages to test.log\n", __FUNCTION__, MAX_TEST4_MESSAGES);
 	fprintf(stderr, "-------------------------------------------------------\n");
+	/*
+	 * generate lua scripts
+	 */
+	assert(chdir(dragonfly_root) == 0);
+	char *path = get_current_dir_name();
+	fprintf(stderr, "DRAGONFLY_ROOT: %s\n", path);
+	free (path);
+	write_file(config_path, CONFIG_LUA);
+	write_file(input_path, INPUT_LUA);
+	write_file(analyzer_path, ANALYZER_LUA);
 
-	const char *test_dir = "/tmp";
-	const char *analyzer_dir = "/tmp/analyzers";
-	const char *script_path = "/tmp/analyzers/test4.lua";
-	const char *config_path = "/tmp/analyzers/config.lua";
-
-	mkdir (analyzer_dir, 0755);
-	// generate a config.lua file 
-	const char* config = "analyzers = {   "
-									 "{ name=\"test4\","
-									   "input=\"ipc:///tmp/test4.ipc\", "
-									   "script=\"test4.lua\", "
-									   "output=\"file:///tmp/test4.log\"}"
-								     "   }\n";
-	FILE *fp = fopen(config_path,"w+");
-	if (!fp)
-	{
-		perror (__FUNCTION__);
-		return;
-	}
-	fprintf (stderr, "config.lua:\n\n%s\n\n", config);
-	fputs(config, fp);
-	fclose (fp);
-	
-	// generate a test.lua analyzer
-	const char* script = "local count = 0 \n"
-					     "function setup()\n"
-						 " count = 0\n"
-						 "end\n"
-					     "function loop (msg)\n"
-						 " count = count + 1\n"
-						 " local now = os.time()\n"
-      					 " local dtg = os.date(\"!%Y-%m-%dT%TZ\",now)\n"
-						 " event = \"event \"..tostring(count)\n"
-						 " log_event (dtg, event, msg)\n"
-						 "end\n\n";
-	fp = fopen(script_path,"w+");
-	if (!fp)
-	{
-		perror (__FUNCTION__);
-		return;
-	}
-	fprintf (stderr, "test4.lua:\n\n%s\n\n", script);
-	fputs(script, fp);
-	fclose (fp);
-	
-
-	// start the service
-	// create a pipe for sending messages
-	// pump messages through the pipe
-	// shutdown the service
 	signal(SIGPIPE, SIG_IGN);
 	openlog("dragonfly", LOG_PERROR, LOG_USER);
 	pthread_setname_np(pthread_self(), "dragonfly");
-    startup_threads(test_dir);
-	sleep (1);
-	
-	DF_HANDLE *pump = dragonfly_io_open("ipc:///tmp/test4.ipc", DF_OUT);
+	startup_threads(dragonfly_root);
+
+	sleep(1);
+
+	DF_HANDLE *pump = dragonfly_io_open("ipc://input.ipc", DF_OUT);
 	if (!pump)
 	{
-		fprintf (stderr, "%s: dragonfly_io_open() failed.\n", __FUNCTION__);
+		fprintf(stderr, "%s: dragonfly_io_open() failed.\n", __FUNCTION__);
 		return;
 	}
 
@@ -121,30 +133,39 @@ void SELF_TEST4(const char *dragonfly_root)
 	 * write messages walking the alphabet
 	 */
 	int mod = 0;
-	for (int i=0; i< MAX_TEST4_MESSAGES; i++)
+	for (unsigned long i = 0; i < MAX_TEST4_MESSAGES; i++)
 	{
-		char msg [128];
-		for (int j=0; j<(sizeof(msg)-1);j++)
+		char msg[128];
+		for (int j = 0; j < (sizeof(msg) - 1); j++)
 		{
-			msg [j]='A'+(mod%48);
+			msg[j] = 'A' + (mod % 48);
 			mod++;
 		}
-		msg [sizeof(msg)-1]='\0';
-    	dragonfly_io_write (pump, msg);
-		if (i==halfway)
+		msg[sizeof(msg) - 1] = '\0';
+		msg[sizeof(msg) - 2] = '\n';
+		if (dragonfly_io_write(pump, msg) < 0)
+		{
+			fprintf(stderr, "error pumping to \"ipc://input.ipc\"\n");
+			abort();
+		}
+		if (i == halfway)
 		{
 			kill(getpid(), SIGUSR1);
 		}
 	}
-	dragonfly_io_close (pump);
+	dragonfly_io_close(pump);
+	sleep(1);
 	shutdown_threads();
-
 	closelog();
-    remove (config_path);
-    remove (script_path);
-	remove ("/tmp/test4.log");
-	rmdir (analyzer_dir);
-	fprintf(stderr,"\n");
+
+	fprintf(stderr, "Cleaning up files\n");
+	remove(config_path);
+	remove(input_path);
+	remove(analyzer_path);	fprintf(stderr, "Cleaning up files\n");
+	remove(config_path);
+	remove(input_path);
+	remove(analyzer_path);
+	fprintf(stderr, "-------------------------------------------------------\n\n");
 }
 
 /*
